add initial docker and pynamodb
