+++ .flake8
@@ -1,0 +1,7 @@
+[flake8]
+max-line-length = 100
+max-complexity = 10
+filename = *.py
+format = default
+
+ignore = E126, E402, E501, E502, E711, E712
