+++ requirements.txt
@@ -1,5 +1,3 @@
-unqlite
-flask
-flask-cors
-Cython
-requests
+flask==0.12.2
+pynamodb==3.1.0
+requests==2.18.1
