+++ ubicajeros/const.py
@@ -1,0 +1,43 @@
+BASE_URL = 'http://www.banxico.org.mx/consultas-atm'
+ORG_NAMES = {
+    37019: 'BANJERCITO',
+    37166: 'BANSEFI',
+    40002: 'BANAMEX',
+    40012: 'BBVA BANCOMER',
+    40014: 'SANTANDER',
+    40021: 'HSBC',
+    40030: 'BAJIO',
+    40036: 'INBURSA',
+    40037: 'INTERACCIONES',
+    40042: 'MIFEL',
+    40044: 'SCOTIABANK',
+    40058: 'BANREGIO',
+    40060: 'BANSI',
+    40062: 'AFIRME',
+    40072: 'BANORTE-IXE',
+    40127: 'AZTECA',
+    40128: 'AUTOFIN',
+    40131: 'BANCO FAMSA',
+    40132: 'BMULTIVA',
+    40134: '',
+    40136: '',
+    40137: 'BANCOPPEL',
+    40138: 'ABC CAPITAL',
+    40143: 'CIBANCO',
+    40147: '',
+    1: 'WAL-MART',
+    2: 'MERCO',
+    3: 'CHEDRAUI',
+    4: 'BODEGA AURRERA',
+    5: 'SAMS',
+    6: 'SUBURBIA',
+    7: 'SUPERAMA'
+}
+STATES = [
+    'AGUASCALIENTES', 'BAJA CALIFORNIA SUR', 'BAJA CALIFORNIA', 'CAMPECHE', 'CHIAPAS',
+    'CHIHUAHUA', 'DISTRITO FEDERAL', 'COAHUILA', 'COLIMA', 'DURANGO', 'GUANAJUATO',
+    'GUERRERO', 'HIDALGO', 'JALISCO', 'ESTADO DE MEXICO', 'MICHOACAN', 'MORELOS',
+    'NAYARIT', 'NUEVO LEON', 'OAXACA', 'PUEBLA', 'QUERETARO', 'QUINTANA ROO',
+    'SAN LUIS POTOSI', 'SINALOA', 'SONORA', 'TABASCO', 'TAMAULIPAS', 'TLAXCALA',
+    'VERACRUZ', 'YUCATAN', 'ZACATECAS'
+]
